Fixeses and tests
fixes:
- temperature data type changed to float64
- removed paralelism for getting telemetry
- added context for ticker

tests:
- utils/byteToInt

+++ temperatures/cpu.go
@@ -2,7 +2,7 @@ package temperatures
 
 import "fmt"
 
-func GetCPUTemperature() (int, error) {
+func GetCPUTemperature() (float64, error) {
 	filepath := telemetryPaths.CPU
 	telemetry, err := getTemperature(filepath)
 	if err != nil {
